Extract duplicate code to a new base class
