+++ plugins/PagePerformance/Columns/TimeTransfer.php
@@ -13,40 +13,15 @@ use Piwik\Columns\MetricsList;
 use Piwik\Piwik;
 use Piwik\Plugin\ArchivedMetric;
 use Piwik\Plugin\ComputedMetric;
-use Piwik\Plugin\Dimension\ActionDimension;
-use Piwik\Tracker\Action;
-use Piwik\Tracker\ActionPageview;
-use Piwik\Tracker\Request;
-use Piwik\Tracker\Visitor;
 
-class TimeTransfer extends ActionDimension
+class TimeTransfer extends Base
 {
     const COLUMN_TYPE = 'MEDIUMINT(10) UNSIGNED NULL';
     const COLUMN_NAME = 'time_transfer';
 
     protected $columnName = self::COLUMN_NAME;
     protected $columnType = self::COLUMN_TYPE;
-    protected $type = self::TYPE_DURATION_MS;
     protected $nameSingular = 'PagePerformance_ColumnTimeTransfer';
-
-    public function onNewAction(Request $request, Visitor $visitor, Action $action)
-    {
-        if (!($action instanceof ActionPageview)) {
-            return false;
-        }
-
-        $transferTime = $request->getParam($this->getRequestParam());
-
-        if ($transferTime === -1) {
-            return false;
-        }
-
-        if ($transferTime < 0) {
-            throw new \Exception(sprintf('Value for %1$s can\'t be negative.', $this->getRequestParam()));
-        }
-
-        return $transferTime;
-    }
 
     public function getRequestParam()
     {
